+++ src/main/kotlin/com/localizedjpa/intellij/migration/ui/LocalizedJpaIcons.kt
@@ -1,0 +1,9 @@
+package com.localizedjpa.intellij.migration.ui
+
+import com.intellij.openapi.util.IconLoader
+import javax.swing.Icon
+
+object LocalizedJpaIcons {
+    @JvmField
+    val GutterIcon: Icon = IconLoader.getIcon("/META-INF/gutterIcon.svg", LocalizedJpaIcons::class.java)
+}